requirements.txt
streamlit
pdfplumber
pytesseract
pillow
lxml
